requirements.txt
streamlit
google-api-python-client
textblob
pandas
plotly
wordcloud
matplotlib
python-dotenv